requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --strip-extras
#
asgiref==3.7.2
    # via django
certifi==2023.7.22
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.2.0
    # via requests
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==41.0.3
    # via pyjwt
defusedxml==0.7.1
    # via
    #   odfpy
    #   python3-openid
diff-match-patch==20230430
    # via django-import-export
django==4.2.5
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-allauth
    #   django-browser-reload
    #   django-crispy-forms
    #   django-htmx
    #   django-import-export
    #   django-template-partials
django-allauth==0.54.0
    # via -r requirements.in
django-browser-reload==1.11.0
    # via -r requirements.in
django-crispy-forms==2.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.10.0
    # via -r requirements.in
django-htmx==1.16.0
    # via -r requirements.in
django-import-export==3.3.1
    # via -r requirements.in
django-template-partials==23.2
    # via -r requirements.in
et-xmlfile==1.1.0
    # via openpyxl
idna==3.4
    # via requests
markuppy==1.14
    # via tablib
oauthlib==3.2.2
    # via requests-oauthlib
odfpy==1.4.1
    # via tablib
openpyxl==3.1.2
    # via tablib
pycparser==2.21
    # via cffi
pyjwt==2.8.0
    # via django-allauth
python3-openid==3.2.0
    # via django-allauth
pyyaml==6.0.1
    # via tablib
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.4
    # via django
tablib==3.5.0
    # via django-import-export
urllib3==2.0.4
    # via requests
xlrd==2.0.1
    # via tablib
xlwt==1.3.0
    # via tablib
